+++ geofront/backends/cloud.py
@@ -9,10 +9,15 @@ with many of the popular cloud service providers using unified API."
 
 """
 import collections.abc
+try:
+    from functools import singledispatch
+except ImportError:
+    from singledispatch import singledispatch
 import io
 import numbers
 
-from libcloud.compute.base import NodeDriver
+from libcloud.compute.base import Node, NodeDriver
+from libcloud.compute.drivers.gce import GCENodeDriver
 from libcloud.storage.base import Container, StorageDriver
 from libcloud.storage.types import ObjectDoesNotExistError
 from paramiko.pkey import PKey
@@ -36,6 +41,11 @@ class CloudRemoteSet(collections.abc.Mapping):
         driver = driver_cls('access id', 'secret key')
         REMOTE_SET = CloudRemoteSet(driver)
 
+    If the given ``driver`` supports metadata feature (for example,
+    AWS EC2, Google Compute Engine, and OpenStack support it)
+    the resulted :class:`~geofront.remote.Remote` objects will
+    fill their :attr:`~geofront.remote.Remote.metadata` as well.
+
     :param driver: libcloud compute driver
     :type driver: :class:`libcloud.compute.base.NodeDriver`
     :param user: the username to :program:`ssh`.
@@ -56,6 +66,10 @@ class CloudRemoteSet(collections.abc.Mapping):
     .. _Libcloud: http://libcloud.apache.org/
     __ https://libcloud.readthedocs.org/en/latest/compute/
 
+    .. versionchanged:: group
+       It fills :attr:`~geofront.remote.Remote.metadata` of the resulted
+       :class:`~geofront.remote.Remote` objects if the ``driver`` supports.
+
     """
 
     @typed
@@ -67,12 +81,15 @@ class CloudRemoteSet(collections.abc.Mapping):
         self.user = user
         self.port = port
         self._nodes = None
+        self._metadata = {} if supports_metadata(driver) else None
 
     def _get_nodes(self, refresh: bool=False) -> dict:
         if refresh or self._nodes is None:
             self._nodes = {node.name: node
                            for node in self.driver.list_nodes()
                            if node.public_ips}
+            if self._metadata is not None:
+                self._metadata.clear()
         return self._nodes
 
     def __len__(self) -> int:
@@ -83,7 +100,37 @@ class CloudRemoteSet(collections.abc.Mapping):
 
     def __getitem__(self, alias: str) -> Remote:
         node = self._get_nodes()[alias]
-        return Remote(self.user, node.public_ips[0], self.port)
+        if self._metadata is None:
+            metadata = {}
+        else:
+            try:
+                metadata = self._metadata[alias]
+            except KeyError:
+                metadata = get_metadata(self.driver, node)
+                self._metadata[alias] = metadata
+        return Remote(self.user, node.public_ips[0], self.port, metadata)
+
+
+@singledispatch
+def supports_metadata(driver: NodeDriver) -> bool:
+    """Whether this drive type supports metadata?"""
+    return callable(getattr(driver, 'ex_get_metadata', None))
+
+
+@singledispatch
+def get_metadata(driver: NodeDriver, node: Node) -> collections.abc.Mapping:
+    return driver.ex_get_metadata(node)
+
+
+@supports_metadata.register(GCENodeDriver)
+def gce_supports_metadata(driver: GCENodeDriver) -> bool:
+    return True
+
+
+@get_metadata.register(GCENodeDriver)
+def gce_get_metadata(driver: GCENodeDriver,
+                     node: Node) -> collections.abc.Mapping:
+    return node.extra['metadata']
 
 
 class CloudMasterKeyStore(MasterKeyStore):